backend/requirements.txt
# Core Backend Dependencies
fastapi==0.109.0
sqlmodel==0.0.14
psycopg[binary]==3.2.3
uvicorn[standard]==0.27.0
python-dotenv==1.0.0

# Authentication Dependencies
PyJWT==2.8.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0  # Pin to 3.2.0 for compatibility with passlib 1.7.4

# Testing Dependencies
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Development Dependencies
black==24.1.1
ruff==0.1.14
mypy==1.8.0
